NMS-632: Add additive search ability for exactUei filter.
The filter builder in EventFilterController is an exclusive
filter. The SQL statement it builds is an "AND"ed list of all
filters listed via the web page. This is problematic, since you
can only specify a single UEI to include, along with a large list
of UEIs to exclude. If you start to filter down the events to show,
you eventually run into the GET header size limit.

We solve these problems by putting all exactUei filters into their
own list, which we then add as an OrFilter to the filters list we
pass through to EventCriteria. We still keep the list of all filters
as a single list since the web UI page does not handle the an "OR"
list.

Add a form to the event list page to allow the user to select additional
UEIs to add to the list of events to show.